refactoring NG's bean folder to be models
